refactor: OffersContext to use fetchSwaps service
