+++ lib/client/contexts/OffersContext.tsx
@@ -2,36 +2,20 @@
 /* eslint-disable unused-imports/no-unused-vars */
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable @typescript-eslint/no-empty-function */
-import { retrieveDataFromTokensArray, decodeConfig } from "../blockchain-utils";
-import { Swap } from "../swap-utils";
 import { useAuthenticatedUser } from "@/lib/client/hooks/useAuthenticatedUser";
-import {
-  ACCEPTED_OFFERS_QUERY,
-  ALL_OFFERS_QUERY,
-  CANCELED_OFFERS_QUERY,
-  CREATED_OFFERS_QUERY,
-  EXPIRED_OFFERS_QUERY,
-  RECEIVED_OFFERS_QUERY,
-} from "@/lib/client/offer-queries";
 import { EthereumAddress, Token, TokenType } from "@/lib/shared/types";
-import { cleanJsonString } from "@/lib/client/utils";
 import { ADDRESS_ZERO } from "@/lib/client/constants";
 import {
   FormattedSwapOfferAssets,
   PopulatedSwapOfferCard,
-  PageParam,
-  RawSwapOfferInterface,
   PageInfo,
   InifniteQueryData,
 } from "@/lib/client/offers-utils";
 import { SwapContext } from "@/lib/client/contexts";
-import { getSwap } from "@/lib/service/getSwap";
+import { fetchSwaps } from "@/lib/service/fetchSwaps";
 import { useInfiniteQuery } from "@tanstack/react-query";
 import React, { Dispatch, useEffect, useState, useContext } from "react";
-import axios from "axios";
-import { isAddress } from "viem";
-import { useAccount, useNetwork } from "wagmi";
-import toast from "react-hot-toast";
+import { useNetwork } from "wagmi";
 
 export enum PonderFilter {
   ALL_OFFERS = "All Offers",
@@ -119,17 +103,7 @@ const DEFAULT_DATA_INFINITE_QUERY: InifniteQueryData = {
   },
 };
 
-// const DEFAULT_DATA_INFINITE_QUERY: InifniteQueryData = {
-//   pages:
-//     [
-//       {
-//         swapOffers: [DEFAULT_SWAP_OFFER],
-//         pageInfo: DEFAULT_PAGE_INFO,
-//       },
-//     ] | undefined,
-// };
 export const OffersContextProvider = ({ children }: any) => {
-  // States and constants
   const [swapOfferToAccept, setSwapOfferToBeAccepted] =
     useState<PopulatedSwapOfferCard | null>(null);
   const [approvedTokensCount, setApprovedTokensCount] = useState(0);
@@ -142,6 +116,7 @@ export const OffersContextProvider = ({ children }: any) => {
   const [offersFilter, setOffersFilter] = useState<PonderFilter>(
     PonderFilter.ALL_OFFERS,
   );
+
   const [tokensList, setTokensList] = useState<PopulatedSwapOfferCard[]>([
     DEFAULT_SWAP_OFFER,
   ]);
@@ -149,227 +124,19 @@ export const OffersContextProvider = ({ children }: any) => {
   const [isLoadingOffersQuery, setIsLoadingOffersQuery] = useState(false);
 
   const { authenticatedUserAddress } = useAuthenticatedUser();
-  const { address, isConnected } = useAccount();
-
   const { destinyChain } = useContext(SwapContext);
   const { chain } = useNetwork();
 
   const userAddress = authenticatedUserAddress?.address;
+  let chainId: number;
 
-  const currentUnixTimeSeconds = Math.floor(new Date().getTime() / 1000);
+  if (typeof chain?.id !== "undefined") {
+    chainId = chain?.id;
+  }
 
   // Functions
   const acceptSwapOffer = async (swap: PopulatedSwapOfferCard) => {
     setSwapOfferToBeAccepted(swap);
-  };
-
-  const fetchSwaps = async ({
-    pageParam,
-  }: PageParam): Promise<{
-    swapOffers: PopulatedSwapOfferCard[];
-    pageInfo: PageInfo;
-  }> => {
-    if (!userAddress) {
-      throw new Error("User address is not defined");
-    }
-    const after = pageParam || null;
-    let query = "";
-    let variables = {};
-    let chainId: number | undefined = undefined;
-
-    if (typeof chain?.id !== "undefined") {
-      chainId = chain?.id;
-    }
-
-    switch (offersFilter) {
-      case PonderFilter.ALL_OFFERS:
-        query = ALL_OFFERS_QUERY;
-        variables = {
-          orderBy: "blockTimestamp",
-          orderDirection: "desc",
-          inputAddress: userAddress,
-          after: after,
-          allowed: userAddress,
-          network: chainId,
-        };
-        break;
-      case PonderFilter.CREATED:
-        query = CREATED_OFFERS_QUERY;
-        variables = {
-          orderBy: "blockTimestamp",
-          orderDirection: "desc",
-          inputAddress: userAddress,
-          after: after,
-          expiry_gt: currentUnixTimeSeconds,
-          network: chainId,
-        };
-        break;
-      case PonderFilter.RECEIVED:
-        query = RECEIVED_OFFERS_QUERY;
-        variables = {
-          orderBy: "blockTimestamp",
-          orderDirection: "desc",
-          after: after,
-          allowed: userAddress,
-          expiry_gt: currentUnixTimeSeconds,
-          network: chainId,
-        };
-        break;
-      case PonderFilter.ACCEPTED:
-        query = ACCEPTED_OFFERS_QUERY;
-        variables = {
-          orderBy: "blockTimestamp",
-          orderDirection: "desc",
-          inputAddress: userAddress,
-          after: after,
-          allowed: userAddress,
-          network: chainId,
-        };
-        break;
-      case PonderFilter.CANCELED:
-        query = CANCELED_OFFERS_QUERY;
-        variables = {
-          orderBy: "blockTimestamp",
-          orderDirection: "desc",
-          inputAddress: userAddress,
-          after: after,
-          allowed: userAddress,
-          network: chainId,
-        };
-        break;
-      case PonderFilter.EXPIRED:
-        query = EXPIRED_OFFERS_QUERY;
-        variables = {
-          orderBy: "blockTimestamp",
-          orderDirection: "desc",
-          inputAddress: userAddress,
-          expiry_lt: currentUnixTimeSeconds,
-          after: after,
-          network: chainId,
-        };
-        break;
-      default:
-        console.error("Invalid offersFilter:", offersFilter);
-        throw new Error("Invalid offersFilter");
-    }
-
-    const endpoint = process.env.NEXT_PUBLIC_PONDER_ENDPOINT;
-    const headers = {
-      "Content-Type": "application/json",
-    };
-
-    if (!endpoint) {
-      throw new Error(
-        "NEXT_PUBLIC_PONDER_ENDPOINT is not defined in the environment variables.",
-      );
-    }
-
-    const findStatus = (swap: FormattedSwapOfferAssets): PonderFilter => {
-      switch (swap.status.toUpperCase()) {
-        case PonderFilter.ACCEPTED.toUpperCase():
-          return PonderFilter.ACCEPTED;
-        case PonderFilter.ALL_OFFERS.toUpperCase():
-          return PonderFilter.ALL_OFFERS;
-        case PonderFilter.CANCELED.toUpperCase():
-          return PonderFilter.CANCELED;
-        case PonderFilter.CREATED.toUpperCase():
-          return PonderFilter.CREATED;
-        case PonderFilter.EXPIRED.toUpperCase():
-          return PonderFilter.EXPIRED;
-        default:
-          return PonderFilter.RECEIVED;
-      }
-    };
-
-    try {
-      const response = await axios.post(
-        endpoint,
-        { query, variables },
-        { headers },
-      );
-      if (response.data && response.data.data) {
-        const items = response.data.data.swapDatabases
-          .items as RawSwapOfferInterface[];
-        const pageInfo = response.data.data.swapDatabases.pageInfo as PageInfo;
-        const processedItems: RawSwapOfferInterface[] = items.map(
-          (obj: any) => {
-            return {
-              ...obj,
-              bid: cleanJsonString(obj.bid),
-              ask: cleanJsonString(obj.ask),
-            };
-          },
-        );
-        const itemsArrayAsSwapOffers: FormattedSwapOfferAssets[] =
-          processedItems.map((item) => {
-            return {
-              id: item.swapId,
-              status: item.status,
-              expiryDate: item.expiry,
-              bidderAssets: {
-                address: isAddress(item.allowed)
-                  ? new EthereumAddress(item.allowed)
-                  : new EthereumAddress(ADDRESS_ZERO),
-                tokens: item.bid,
-              },
-              askerAssets: {
-                address: new EthereumAddress(item.owner),
-                tokens: item.ask,
-              },
-            };
-          });
-        const itemsArrayAsSwapOffersPopulated: Promise<PopulatedSwapOfferCard>[] =
-          itemsArrayAsSwapOffers.map(async (swap) => {
-            const bidedTokensWithData = await retrieveDataFromTokensArray(
-              swap.bidderAssets.tokens,
-            );
-            const askedTokensWithData = await retrieveDataFromTokensArray(
-              swap.askerAssets.tokens,
-            );
-            const swapStatus = findStatus(swap);
-            const swapData: Swap = await getSwap(
-              BigInt(swap.id),
-              chainId as number,
-            );
-            const swapExpiryData = await decodeConfig({
-              config: swapData.config,
-            });
-
-            return {
-              id: BigInt(swap.id),
-              status: swapStatus,
-              expiryDate: swapExpiryData.expiry,
-              bidderTokens: {
-                address: swap.bidderAssets.address,
-                tokens: bidedTokensWithData,
-              },
-              askerTokens: {
-                address: swap.askerAssets.address,
-                tokens: askedTokensWithData,
-              },
-            };
-          });
-
-        const formattedTokens: PopulatedSwapOfferCard[] = await Promise.all(
-          itemsArrayAsSwapOffersPopulated,
-        );
-        setOffersQueries({
-          ...offersQueries,
-          [offersFilter]: formattedTokens,
-        });
-        return {
-          swapOffers: formattedTokens,
-          pageInfo,
-        };
-      } else {
-        throw new Error("Failed to fetch swaps from Subgraph");
-      }
-    } catch (error) {
-      toast.error(
-        "Failed to fetch swaps from Subgraph. Please contact the team",
-      );
-      throw new Error("Failed to fetch swaps from Subgraph");
-    }
   };
 
   // Offers query
@@ -382,7 +149,12 @@ export const OffersContextProvider = ({ children }: any) => {
         destinyChain,
       ],
       queryFn: async ({ pageParam }: { pageParam: string | null }) =>
-        await fetchSwaps({ pageParam }),
+        await fetchSwaps({
+          pageParam: pageParam,
+          userAddress: userAddress,
+          offersFilter: offersFilter,
+          chainId: chainId,
+        }),
       initialPageParam: null,
       refetchOnWindowFocus: false,
       staleTime: Infinity,
@@ -409,8 +181,6 @@ export const OffersContextProvider = ({ children }: any) => {
         }));
       },
     });
-
-  console.log("data,", data);
 
   useEffect(() => {
     refetch();
